[WIP] Adapt recar to the unminified version of Racer
The Racer website has changed. The scripts are now unminified, and some
pieces of code have been commented by the devs for some reason, breaking
some functionality.

This change adapts the download script so it continues working, along
with a new patch file (which is based in the old one with added
patches).

Fixes: #1